Fixed black spots issue
When the initial hit on a mesh results in a ray origin
behind the surface there is a stronger chance of the next
ray direction escaping into infinity. This results in
darker spots on specific places in the final image.
The applied solution starts a random ray direction at
a small offset along the new ray direction. This skips
any intersections with the origin surface.